pyproject.toml
[project]
name = "tttd"
version = "0.1.0"
description = "Test-Time Training on Erdős Minimum Overlap Problem"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chz>=0.4.0",
    "numpy",
    "python-dotenv",
    "tinker>=0.9.0",
    "tinker-cookbook",
]

[project.scripts]
tttd = "tttd.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/tttd"]

[tool.uv.sources]
tinker-cookbook = { git = "https://github.com/thinking-machines-lab/tinker-cookbook.git" }
